Split repositories.py: three aggregates that never shared anything
1,117 lines holding runtime profiles, agents and conversations — three things
with no relationship to each other behind one import. Now four files, largest
527:

  runtime_profile_repository.py      260
  agent_repository.py                243
  conversation_repository.py         527
  conversation_approval_queries.py   143

The last one is not arbitrary. `ConversationRepository` alone is 577 lines of
class, which lands over 600 once a file has imports — and the ratchet keys
`oversized_files` by *path*, so a new file has an implicit baseline of zero and
620 would have failed CI even though the total shrank. The five approval reads
and writes are a coherent group on their own — what did the person decide, and
which call was it about — and they follow the `ConversationRunQueriesMixin`
precedent that already lives beside them.

The package name is unchanged, so all 37 importers still say
`from app.modules.agent.infrastructure.repositories import X` and none of them
moved. Exactly three names are imported anywhere; `__init__` re-exports those
three and nothing else.

Caught before it shipped this time: copying the shared preamble into each new
file left thirteen module constants in files that never read them — the same
bug the agent_surfaces split shipped and had reported back by the code-quality
bot. Ruff prunes unused imports but not unused assignments, so the check has to
be explicit. There are none.

5,329 unit tests, 114 agent e2e tests, `typecheck-critical` and the import
budget all pass. agent: 6 oversized files -> 5.
