--- lemma-backend/app/modules/agent/infrastructure/repositories.py
@@ -1,1117 +1,0 @@
-"""Repositories for the unified agent module."""
-
-from __future__ import annotations
-
-from collections.abc import Sequence
-from datetime import datetime, timezone
-from uuid import UUID
-
-from pydantic import ValidationError
-from sqlalchemy import func, literal, literal_column, select, update
-from sqlalchemy.dialects.postgresql import JSONB, array
-from sqlalchemy.orm import selectinload
-
-from app.core.authorization.context import Context, ResourceType, ResourceVisibility
-from app.core.log.log import get_logger
-from app.core.authorization.delegation import DEFAULT_POD_AGENT_ID
-from app.core.authorization.grants import (
-    delete_grantee_grants,
-    delete_resource_grants,
-    delete_resource_sharing_grants,
-)
-from app.core.authorization.permissions import Permissions
-from app.core.authorization.sql_actions import (
-    allowed_actions_contains,
-    allowed_actions_expr,
-)
-from app.core.infrastructure.db.uow import SqlAlchemyUnitOfWork
-from app.modules.agent.domain.events import (
-    AgentCreatedEvent,
-    AgentDomainEvent,
-    ConversationStartedEvent,
-)
-from app.modules.agent.domain.entities import (
-    Agent as AgentEntity,
-    AgentRun as AgentRunEntity,
-    Conversation as ConversationEntity,
-    Message as MessageEntity,
-)
-from app.modules.agent.domain.runtime_profiles import (
-    AgentRuntimeProfile,
-    RuntimeProfileProtocol,
-    RuntimeProfileScope,
-    RuntimeProfileStatus,
-    reveal_credentials,
-)
-from app.modules.agent.domain.value_objects import (
-    ACTIVE_AGENT_RUN_STATUSES,
-    TERMINAL_AGENT_RUN_STATUSES,
-    AgentRuntimeConfig,
-    AgentRunApprovalDecision,
-    AgentRunFinishResult,
-    AgentRunStatus,
-    ConversationAgentScope,
-    ConversationAgentSelection,
-    ConversationStatus,
-    ConversationType,
-    JsonObject,
-    JsonValue,
-    MessageDraft,
-    MessageKind,
-    to_json_value,
-)
-from app.modules.agent.infrastructure.models import (
-    AgentApprovalDecisionModel,
-    AgentModel,
-    AgentRunModel,
-    ConversationModel,
-    MessageModel,
-)
-from app.modules.agent.infrastructure.runtime_models import (
-    AgentRuntimeProfileModel,
-)
-from app.modules.agent.infrastructure.conversation_origin_store import (
-    create_conversation_for_origin,
-)
-from app.modules.agent.infrastructure.conversation_run_queries import (
-    ConversationRunQueriesMixin,
-)
-from app.modules.agent.infrastructure.repository_status import (
-    conversation_status_values_for_db as _conversation_status_values_for_db,
-    run_status_values_for_db as _run_status_values_for_db,
-)
-from app.modules.connectors.contracts import SecretEncryptionPort
-
-
-_ACTIVE_AGENT_RUN_STATUS_VALUES = _run_status_values_for_db(ACTIVE_AGENT_RUN_STATUSES)
-_DEFAULT_POD_AGENT_ID_SQL = literal_column(f"'{DEFAULT_POD_AGENT_ID}'::uuid")
-
-logger = get_logger(__name__)
-
-# Rows written for the retired local daemon still carry its per-tool protocols
-# (CODEX_APP_SERVER, CLAUDE_CODE, …). They can never execute again, and
-# RuntimeProfileProtocol no longer parses them, so every read filters on the
-# protocols still understood rather than letting one stale row fail the query
-# for the whole organization.
-_KNOWN_RUNTIME_PROFILE_PROTOCOLS = [
-    protocol.value for protocol in RuntimeProfileProtocol
-]
-
-
-class AgentRuntimeProfileRepository:
-    """Repository for organization-owned runtime profiles."""
-
-    def __init__(
-        self,
-        uow: SqlAlchemyUnitOfWork,
-        encryption: SecretEncryptionPort,
-    ):
-        self.uow = uow
-        self.session = uow.session
-        self.encryption = encryption
-
-    @staticmethod
-    def _serialize_json(value: object | None) -> dict | None:
-        if value is None:
-            return None
-        model_dump = getattr(value, "model_dump", None)
-        if callable(model_dump):
-            return model_dump(mode="json")
-        if isinstance(value, dict):
-            return value
-        return None
-
-    def _to_entity(self, instance: AgentRuntimeProfileModel) -> AgentRuntimeProfile:
-        data = instance.to_entity().model_dump(mode="json")
-        data["credentials"] = self.encryption.decrypt_json(instance.credentials)
-        return AgentRuntimeProfile.model_validate(data)
-
-    def _apply_entity(
-        self,
-        instance: AgentRuntimeProfileModel,
-        entity: AgentRuntimeProfile,
-    ) -> AgentRuntimeProfileModel:
-        """Write the columns an edit may change.
-
-        Identity - organization, user, scope, kind, runtime_type, harness_id -
-        is deliberately absent. runtime_type and harness_id are bound together
-        by ck_agent_runtime_profile_harness_binding, and changing scope would
-        move the row between the two partial unique indexes, so both belong to
-        creation only.
-        """
-        instance.name = entity.name
-        instance.description = entity.description
-        instance.default_model_name = entity.default_model_name
-        # Fresh containers rather than in-place mutation, so SQLAlchemy sees the
-        # JSONB attributes as dirty and actually writes them.
-        instance.model_catalog = [
-            item.model_dump(mode="json") for item in entity.model_catalog
-        ]
-        instance.config = self._serialize_json(entity.config) or {}
-        # reveal_credentials (not _serialize_json) so SecretStr fields persist
-        # as their plaintext before encryption — model_dump(mode="json") would
-        # otherwise store the masked "**********" and lose the real key.
-        instance.credentials = self.encryption.encrypt_json(
-            reveal_credentials(entity.credentials)
-        )
-        instance.status = entity.status.value
-        instance.profile_metadata = dict(entity.metadata)
-        return instance
-
-    def _to_model(self, entity: AgentRuntimeProfile) -> AgentRuntimeProfileModel:
-        return self._apply_entity(
-            AgentRuntimeProfileModel(
-                organization_id=entity.organization_id,
-                user_id=entity.user_id,
-                harness_id=entity.harness_id,
-                # runtime_type mirrors kind: the harness-binding check constraint
-                # is expressed over runtime_type, not kind.
-                runtime_type=entity.kind.value,
-                scope=entity.scope.value,
-                kind=entity.kind.value,
-                protocol=entity.protocol.value,
-            ),
-            entity,
-        )
-
-    async def create(self, entity: AgentRuntimeProfile) -> AgentRuntimeProfile:
-        instance = self._to_model(entity)
-        self.session.add(instance)
-        await self._flush_unique(entity.name)
-        return self._to_entity(instance)
-
-    async def _flush_unique(self, name: str) -> None:
-        """Flush, mapping a name collision to the controller's 409."""
-        from sqlalchemy.exc import IntegrityError
-
-        try:
-            await self.session.flush()
-        except IntegrityError as exc:
-            raise RuntimeError(
-                f"Runtime profile named {name!r} already exists"
-            ) from exc
-
-    async def update(self, entity: AgentRuntimeProfile) -> AgentRuntimeProfile:
-        instance = await self._load(profile_id=entity.id)
-        if instance is None:
-            raise RuntimeError(f"Runtime profile {entity.id!r} no longer exists")
-        self._apply_entity(instance, entity)
-        await self._flush_unique(entity.name)
-        return self._to_entity(instance)
-
-    async def set_status(
-        self,
-        *,
-        profile_id: str,
-        organization_id: UUID,
-        user_id: UUID,
-        status: RuntimeProfileStatus,
-    ) -> AgentRuntimeProfile | None:
-        instance = await self._load(
-            profile_id=profile_id,
-            organization_id=organization_id,
-            user_id=user_id,
-            # An already-archived profile has to be loadable, or it could never
-            # be restored.
-            include_disabled=True,
-        )
-        if instance is None:
-            return None
-        instance.status = status.value
-        # Restoring can collide: another profile may have taken this name while
-        # this one sat archived.
-        await self._flush_unique(instance.name)
-        return self._to_entity(instance)
-
-    async def _load(
-        self,
-        *,
-        profile_id: str,
-        organization_id: UUID | None = None,
-        user_id: UUID | None = None,
-        include_disabled: bool = False,
-    ) -> AgentRuntimeProfileModel | None:
-        try:
-            profile_uuid = UUID(profile_id)
-        except ValueError:
-            return None
-        stmt = select(AgentRuntimeProfileModel).where(
-            AgentRuntimeProfileModel.id == profile_uuid,
-            AgentRuntimeProfileModel.protocol.in_(_KNOWN_RUNTIME_PROFILE_PROTOCOLS),
-        )
-        if organization_id is not None:
-            stmt = stmt.where(
-                AgentRuntimeProfileModel.organization_id == organization_id
-            )
-        if user_id is not None:
-            stmt = stmt.where(
-                (
-                    AgentRuntimeProfileModel.scope
-                    == RuntimeProfileScope.ORGANIZATION.value
-                )
-                | (
-                    (
-                        AgentRuntimeProfileModel.scope
-                        == RuntimeProfileScope.PERSONAL.value
-                    )
-                    & (AgentRuntimeProfileModel.user_id == user_id)
-                )
-            )
-        if not include_disabled:
-            stmt = stmt.where(
-                AgentRuntimeProfileModel.status == RuntimeProfileStatus.ACTIVE.value
-            )
-        result = await self.session.execute(stmt.limit(1))
-        return result.scalar_one_or_none()
-
-    async def get_visible(
-        self,
-        *,
-        organization_id: UUID,
-        user_id: UUID,
-        include_disabled: bool = False,
-    ) -> list[AgentRuntimeProfile]:
-        stmt = select(AgentRuntimeProfileModel).where(
-            AgentRuntimeProfileModel.organization_id == organization_id,
-            AgentRuntimeProfileModel.protocol.in_(_KNOWN_RUNTIME_PROFILE_PROTOCOLS),
-            (AgentRuntimeProfileModel.scope == RuntimeProfileScope.ORGANIZATION.value)
-            | (
-                (AgentRuntimeProfileModel.scope == RuntimeProfileScope.PERSONAL.value)
-                & (AgentRuntimeProfileModel.user_id == user_id)
-            ),
-        )
-        if not include_disabled:
-            stmt = stmt.where(
-                AgentRuntimeProfileModel.status == RuntimeProfileStatus.ACTIVE.value
-            )
-        stmt = stmt.order_by(
-            AgentRuntimeProfileModel.name.asc(),
-        )
-        result = await self.session.execute(stmt)
-        # One unmappable row must not blank the workspace's entire model list.
-        # The protocol filter above already excludes retired daemon profiles, so
-        # anything reaching here has a protocol this build knows and a body it
-        # could not validate - a partial write, a hand-edited row, or a field
-        # this build tightened. Skipping it costs that one profile; raising
-        # costs every profile the organization has, and the Models page with it.
-        profiles: list[AgentRuntimeProfile] = []
-        for instance in result.scalars():
-            try:
-                profiles.append(self._to_entity(instance))
-            except ValidationError as error:
-                logger.warning(
-                    "agent.runtime_profile.unreadable.skipped",
-                    profile_id=str(instance.id),
-                    organization_id=str(organization_id),
-                    error=str(error),
-                )
-        return profiles
-
-    async def get_visible_by_id(
-        self,
-        *,
-        profile_id: str,
-        organization_id: UUID,
-        user_id: UUID,
-        include_disabled: bool = False,
-    ) -> AgentRuntimeProfile | None:
-        instance = await self._load(
-            profile_id=profile_id,
-            organization_id=organization_id,
-            user_id=user_id,
-            include_disabled=include_disabled,
-        )
-        return self._to_entity(instance) if instance else None
-
-
-class AgentRepository:
-    """Repository for pod-owned agents."""
-
-    def __init__(self, uow: SqlAlchemyUnitOfWork):
-        self.uow = uow
-        self.session = uow.session
-
-    async def create(self, agent: AgentEntity) -> AgentEntity:
-        model = AgentModel(
-            id=agent.id,
-            created_at=agent.created_at,
-            updated_at=agent.updated_at,
-            pod_id=agent.pod_id,
-            user_id=agent.user_id,
-            name=agent.name,
-            description=agent.description,
-            icon_url=agent.icon_url,
-            visibility=agent.visibility,
-            instruction=agent.instruction,
-            agent_runtime=(
-                agent.agent_runtime.model_dump(mode="json")
-                if agent.agent_runtime
-                else None
-            ),
-            toolsets=[toolset.value for toolset in agent.toolsets],
-            input_schema=agent.input_schema,
-            output_schema=agent.output_schema,
-            agent_metadata=agent.metadata,
-        )
-        self.session.add(model)
-        await self.session.flush()
-        # The single write path behind every creation route.
-        self.uow.collect_events(
-            [
-                AgentCreatedEvent(
-                    agent_id=model.id,
-                    pod_id=model.pod_id,
-                    user_id=model.user_id,
-                    tool_count=len(agent.toolsets or ()),
-                )
-            ]
-        )
-        return model.to_entity()
-
-    def _to_entity_with_allowed_actions(
-        self,
-        model: AgentModel,
-        allowed_actions: list[str] | tuple[str, ...] | None = None,
-    ) -> AgentEntity:
-        entity = model.to_entity()
-        if allowed_actions is not None:
-            entity.allowed_actions = list(allowed_actions)
-        return entity
-
-    async def get(
-        self, agent_id: UUID, ctx: Context | None = None
-    ) -> AgentEntity | None:
-        if ctx is None:
-            result = await self.session.execute(
-                select(AgentModel).where(AgentModel.id == agent_id)
-            )
-            model = result.scalar_one_or_none()
-            return model.to_entity() if model else None
-        actions = allowed_actions_expr(
-            ctx=ctx,
-            resource_type=ResourceType.AGENT,
-            resource_id_col=AgentModel.id,
-            pod_id_col=AgentModel.pod_id,
-            owner_user_id_col=AgentModel.user_id,
-            visibility_col=AgentModel.visibility,
-        )
-        result = await self.session.execute(
-            select(AgentModel, actions).where(AgentModel.id == agent_id)
-        )
-        row = result.one_or_none()
-        return self._to_entity_with_allowed_actions(row[0], row[1]) if row else None
-
-    async def update(self, agent: AgentEntity) -> AgentEntity:
-        result = await self.session.execute(
-            select(AgentModel).where(AgentModel.id == agent.id)
-        )
-        model = result.scalar_one_or_none()
-        if model is None:
-            return agent
-
-        model.name = agent.name
-        model.description = agent.description
-        model.icon_url = agent.icon_url
-        previous_visibility = model.visibility
-        model.visibility = agent.visibility
-        if (
-            previous_visibility == ResourceVisibility.RESTRICTED.value
-            and agent.visibility != ResourceVisibility.RESTRICTED.value
-        ):
-            await delete_resource_sharing_grants(
-                self.session,
-                pod_id=agent.pod_id,
-                resource_type=ResourceType.AGENT,
-                resource_id=agent.id,
-            )
-        model.instruction = agent.instruction
-        model.agent_runtime = (
-            agent.agent_runtime.model_dump(mode="json") if agent.agent_runtime else None
-        )
-        model.toolsets = [toolset.value for toolset in agent.toolsets]
-        model.input_schema = agent.input_schema
-        model.output_schema = agent.output_schema
-        model.agent_metadata = agent.metadata
-        await self.session.flush()
-        return model.to_entity()
-
-    async def delete(self, agent_id: UUID) -> None:
-        result = await self.session.execute(
-            select(AgentModel).where(AgentModel.id == agent_id)
-        )
-        model = result.scalar_one_or_none()
-        if model is not None:
-            if model.pod_id is not None:
-                await delete_resource_grants(
-                    self.session,
-                    pod_id=model.pod_id,
-                    resource_type=ResourceType.AGENT,
-                    resource_id=agent_id,
-                )
-                await delete_grantee_grants(
-                    self.session,
-                    pod_id=model.pod_id,
-                    grantee_type="AGENT",
-                    grantee_id=agent_id,
-                )
-            await self.session.delete(model)
-            await self.session.flush()
-
-    async def list_by_pod(
-        self,
-        *,
-        pod_id: UUID,
-        cursor: UUID | None = None,
-        limit: int = 100,
-    ) -> tuple[list[AgentEntity], UUID | None]:
-        stmt = select(AgentModel).where(AgentModel.pod_id == pod_id)
-        if cursor is not None:
-            stmt = stmt.where(AgentModel.id < cursor)
-        stmt = stmt.order_by(AgentModel.id.desc()).limit(limit + 1)
-        result = await self.session.execute(stmt)
-        rows = list(result.scalars())
-        has_more = len(rows) > limit
-        if has_more:
-            rows = rows[:limit]
-        next_cursor = rows[-1].id if has_more and rows else None
-        return [row.to_entity() for row in rows], next_cursor
-
-    async def list_visible_by_pod(
-        self,
-        *,
-        pod_id: UUID,
-        ctx: Context,
-        cursor: UUID | None = None,
-        limit: int = 100,
-    ) -> tuple[list[AgentEntity], UUID | None]:
-        actions = allowed_actions_expr(
-            ctx=ctx,
-            resource_type=ResourceType.AGENT,
-            resource_id_col=AgentModel.id,
-            pod_id_col=AgentModel.pod_id,
-            owner_user_id_col=AgentModel.user_id,
-            visibility_col=AgentModel.visibility,
-        )
-        stmt = select(AgentModel, actions).where(
-            AgentModel.pod_id == pod_id,
-            allowed_actions_contains(actions, Permissions.AGENT_READ),
-        )
-        if cursor is not None:
-            stmt = stmt.where(AgentModel.id < cursor)
-        stmt = stmt.order_by(AgentModel.id.desc()).limit(limit + 1)
-        result = await self.session.execute(stmt)
-        rows = list(result.all())
-        has_more = len(rows) > limit
-        if has_more:
-            rows = rows[:limit]
-        next_cursor = rows[-1][0].id if has_more and rows else None
-        return [
-            self._to_entity_with_allowed_actions(model, actions)
-            for model, actions in rows
-        ], next_cursor
-
-    async def get_by_pod_and_name(
-        self, *, pod_id: UUID, name: str, ctx: Context | None = None
-    ) -> AgentEntity | None:
-        if ctx is None:
-            result = await self.session.execute(
-                select(AgentModel).where(
-                    AgentModel.pod_id == pod_id, AgentModel.name == name
-                )
-            )
-            model = result.scalar_one_or_none()
-            return model.to_entity() if model else None
-        actions = allowed_actions_expr(
-            ctx=ctx,
-            resource_type=ResourceType.AGENT,
-            resource_id_col=AgentModel.id,
-            pod_id_col=AgentModel.pod_id,
-            owner_user_id_col=AgentModel.user_id,
-            visibility_col=AgentModel.visibility,
-        )
-        result = await self.session.execute(
-            select(AgentModel, actions).where(
-                AgentModel.pod_id == pod_id, AgentModel.name == name
-            )
-        )
-        row = result.one_or_none()
-        return self._to_entity_with_allowed_actions(row[0], row[1]) if row else None
-
-
-class ConversationRepository(ConversationRunQueriesMixin):
-    """Repository for conversations, agent runs, and messages."""
-
-    def __init__(self, uow: SqlAlchemyUnitOfWork):
-        self.uow = uow
-        self.session = uow.session
-
-    def collect_events(self, events: Sequence[AgentDomainEvent]) -> None:
-        self.uow.collect_events(events)
-
-    async def create_conversation(
-        self,
-        conversation: ConversationEntity,
-    ) -> ConversationEntity:
-        model = ConversationModel(
-            id=conversation.id,
-            created_at=conversation.created_at,
-            updated_at=conversation.updated_at,
-            user_id=conversation.user_id,
-            pod_id=conversation.pod_id,
-            organization_id=conversation.organization_id,
-            agent_id=conversation.agent_id,
-            title=conversation.title,
-            instructions=conversation.instructions,
-            agent_runtime=(
-                conversation.agent_runtime.model_dump(mode="json")
-                if conversation.agent_runtime
-                else None
-            ),
-            origin_type=conversation.origin_type,
-            origin_id=conversation.origin_id,
-            conversation_type=conversation.type.value,
-            status=conversation.status.value if conversation.status else None,
-            output_data=conversation.output,
-            parent_id=conversation.parent_id,
-            conversation_metadata=conversation.metadata,
-        )
-        self.session.add(model)
-        await self.session.flush()
-        self.uow.collect_events(
-            [
-                ConversationStartedEvent(
-                    conversation_id=model.id,
-                    pod_id=model.pod_id,
-                    user_id=model.user_id,
-                    agent_id=model.agent_id,
-                    parent_id=model.parent_id,
-                )
-            ]
-        )
-        return model.to_entity()
-
-    async def create_conversation_once(
-        self,
-        conversation: ConversationEntity,
-    ) -> tuple[ConversationEntity, bool]:
-        """Create one conversation for a durable external invocation origin."""
-        return await create_conversation_for_origin(self.session, conversation)
-
-    async def update_conversation(
-        self,
-        conversation: ConversationEntity,
-    ) -> ConversationEntity:
-        result = await self.session.execute(
-            select(ConversationModel).where(ConversationModel.id == conversation.id)
-        )
-        model = result.scalar_one_or_none()
-        if model is None:
-            return conversation
-
-        model.title = conversation.title
-        model.instructions = conversation.instructions
-        model.agent_runtime = (
-            conversation.agent_runtime.model_dump(mode="json")
-            if conversation.agent_runtime
-            else None
-        )
-        model.conversation_type = conversation.type.value
-        model.status = conversation.status.value if conversation.status else None
-        model.output_data = conversation.output
-        model.conversation_metadata = conversation.metadata
-        await self.session.flush()
-        return model.to_entity()
-
-    async def get_conversation_metadata_key(
-        self,
-        conversation_id: UUID,
-        key: str,
-    ) -> JsonValue | None:
-        """Read a single key out of a conversation's metadata JSON blob."""
-        result = await self.session.execute(
-            select(ConversationModel.conversation_metadata).where(
-                ConversationModel.id == conversation_id
-            )
-        )
-        metadata = result.scalar_one_or_none()
-        if not isinstance(metadata, dict):
-            return None
-        return metadata.get(key)
-
-    async def set_conversation_metadata_key(
-        self,
-        conversation_id: UUID,
-        key: str,
-        value: JsonValue,
-    ) -> None:
-        """Write a single metadata key without clobbering sibling keys.
-
-        Uses ``jsonb_set`` so concurrent writers touching other keys (e.g. the
-        ``is_sub_agent`` / ``surface_platform`` flags) are never overwritten.
-
-        ``value`` must be bound with the ``JSONB`` type directly (``literal``,
-        not ``cast(json.dumps(value), JSONB)``) — casting an already-serialized
-        JSON string double-encodes it, so ``jsonb_set`` stores a JSON *string*
-        scalar (the dumped text) instead of the intended array/object.
-        """
-        stmt = (
-            update(ConversationModel)
-            .where(ConversationModel.id == conversation_id)
-            .values(
-                conversation_metadata=func.jsonb_set(
-                    func.coalesce(
-                        ConversationModel.conversation_metadata,
-                        literal({}, JSONB),
-                    ),
-                    array([key]),
-                    literal(value, JSONB),
-                    True,
-                )
-            )
-        )
-        await self.session.execute(stmt)
-        await self.session.flush()
-
-    async def get_conversation(
-        self,
-        conversation_id: UUID,
-        *,
-        include_messages: bool = False,
-        include_runs: bool = False,
-    ) -> ConversationEntity | None:
-        """Load a conversation, optionally with its messages and its latest run.
-
-        ``include_runs`` used to eager-load every run of the conversation and
-        every message of every run. Nothing consumed that: the entity derives
-        four scalars from ``agent_runs[-1]``, and the three callers that pass
-        this flag read the latest run or nothing at all. On a long-lived thread
-        it meant re-reading the whole transcript — the worst conversation in
-        production carries 13,688 messages — to answer a question about one row.
-
-        Now it fetches exactly that one run, through the index that already
-        exists for it. ``agent_runs`` still carries a single element so
-        ``[-1]`` keeps working for callers, and ``to_entity()`` derives
-        ``status`` and the ``last_run_*`` fields from it unchanged.
-        """
-        stmt = select(ConversationModel).where(ConversationModel.id == conversation_id)
-        if include_messages:
-            stmt = stmt.options(selectinload(ConversationModel.messages))
-        result = await self.session.execute(stmt)
-        model = result.scalar_one_or_none()
-        if model is None:
-            return None
-        if include_runs:
-            latest = await self.session.scalar(
-                select(AgentRunModel)
-                .where(AgentRunModel.conversation_id == conversation_id)
-                .order_by(AgentRunModel.created_at.desc(), AgentRunModel.id.desc())
-                .limit(1)
-            )
-            # Seeded into __dict__ rather than assigned, so SQLAlchemy does not
-            # treat this as a mutation of the relationship and try to flush it.
-            model.__dict__["agent_runs"] = [latest] if latest is not None else []
-        return model.to_entity()
-
-    async def list_conversations(
-        self,
-        *,
-        user_id: UUID,
-        pod_id: UUID,
-        agent_selection: ConversationAgentSelection[UUID],
-        status: ConversationStatus | None = None,
-        conversation_type: ConversationType | None = None,
-        metadata_filters: JsonObject | None = None,
-        parent_id: UUID | None = None,
-        cursor: UUID | None = None,
-        limit: int = 20,
-    ) -> tuple[list[ConversationEntity], UUID | None]:
-        stmt = select(ConversationModel).where(
-            ConversationModel.user_id == user_id,
-            ConversationModel.pod_id == pod_id,
-        )
-        # Default: root conversations only. With parent_id: that conversation's
-        # children (sub-agent conversations).
-        if parent_id is None:
-            stmt = stmt.where(ConversationModel.parent_id.is_(None))
-        else:
-            stmt = stmt.where(ConversationModel.parent_id == parent_id)
-        if agent_selection.scope is not ConversationAgentScope.ALL:
-            selected_agent_id = DEFAULT_POD_AGENT_ID
-            if agent_selection.scope is ConversationAgentScope.NAMED:
-                selected_agent_id = agent_selection.named_value
-            agent_scope_id = func.coalesce(
-                ConversationModel.agent_id, _DEFAULT_POD_AGENT_ID_SQL
-            )
-            stmt = stmt.where(agent_scope_id == selected_agent_id)
-        if status is not None:
-            stmt = stmt.where(
-                ConversationModel.status.in_(_conversation_status_values_for_db(status))
-            )
-        if conversation_type is not None:
-            stmt = stmt.where(
-                ConversationModel.conversation_type == conversation_type.value
-            )
-        if metadata_filters:
-            stmt = stmt.where(
-                ConversationModel.conversation_metadata.op("@>")(metadata_filters)
-            )
-        return await self._list_conversations(stmt, cursor=cursor, limit=limit)
-
-    async def _list_conversations(
-        self,
-        stmt,
-        *,
-        cursor: UUID | None,
-        limit: int,
-    ) -> tuple[list[ConversationEntity], UUID | None]:
-        if cursor is not None:
-            stmt = stmt.where(ConversationModel.id < cursor)
-        stmt = stmt.order_by(ConversationModel.id.desc()).limit(limit + 1)
-        result = await self.session.execute(stmt)
-        rows = list(result.scalars())
-        has_more = len(rows) > limit
-        if has_more:
-            rows = rows[:limit]
-        next_cursor = rows[-1].id if has_more and rows else None
-        return [row.to_entity() for row in rows], next_cursor
-
-    async def lock_conversation(self, conversation_id: UUID) -> None:
-        await self.session.execute(
-            select(ConversationModel.id)
-            .where(ConversationModel.id == conversation_id)
-            .with_for_update()
-        )
-
-    async def list_children(
-        self,
-        *,
-        parent_id: UUID,
-        user_id: UUID,
-        limit: int = 50,
-        include_runs: bool = True,
-    ) -> list[ConversationEntity]:
-        """List child (sub-agent) conversations of a parent, newest first.
-
-        Inverse of list_conversations (which hides children via parent_id IS NULL);
-        reuses the ix_agent_conv_parent index. Scoped to the owning user.
-
-        ``include_runs`` attaches each child's latest run only. The caller reads
-        ``agent_runs[-1].status`` and nothing else, so eager-loading the full
-        collection meant every run of every child to answer one question per
-        child.
-        """
-        stmt = (
-            select(ConversationModel)
-            .where(
-                ConversationModel.parent_id == parent_id,
-                ConversationModel.user_id == user_id,
-            )
-            .order_by(
-                ConversationModel.created_at.desc(),
-                ConversationModel.id.desc(),
-            )
-            .limit(limit)
-        )
-        result = await self.session.execute(stmt)
-        models = list(result.scalars())
-        if include_runs and models:
-            latest_by_conversation = await self._latest_runs_for(
-                [model.id for model in models]
-            )
-            for model in models:
-                latest = latest_by_conversation.get(model.id)
-                model.__dict__["agent_runs"] = [latest] if latest is not None else []
-        return [model.to_entity() for model in models]
-
-    async def create_agent_run(
-        self,
-        *,
-        conversation_id: UUID,
-        agent_id: UUID | None,
-        agent_runtime: AgentRuntimeConfig,
-        parent_run_id: UUID | None = None,
-        metadata: JsonObject | None = None,
-    ) -> AgentRunEntity:
-        now = datetime.now(timezone.utc)
-        model = AgentRunModel(
-            conversation_id=conversation_id,
-            agent_id=agent_id,
-            parent_run_id=parent_run_id,
-            status=AgentRunStatus.RUNNING.value,
-            agent_runtime=agent_runtime.model_dump(mode="json"),
-            started_at=now,
-            run_metadata=metadata,
-        )
-        self.session.add(model)
-        await self._update_conversation_status(
-            conversation_id=conversation_id,
-            status=ConversationStatus.RUNNING,
-            output_data=None,
-        )
-        await self.session.flush()
-        return model.to_entity()
-
-    async def append_message(
-        self,
-        *,
-        conversation_id: UUID,
-        agent_run_id: UUID | None,
-        draft: MessageDraft,
-    ) -> MessageEntity:
-        lock_result = await self.session.execute(
-            select(ConversationModel)
-            .where(ConversationModel.id == conversation_id)
-            .with_for_update()
-        )
-        conversation = lock_result.scalar_one()
-        sequence_result = await self.session.execute(
-            select(func.coalesce(func.max(MessageModel.sequence), -1)).where(
-                MessageModel.conversation_id == conversation_id
-            )
-        )
-        sequence = int(sequence_result.scalar_one()) + 1
-        model = MessageModel(
-            conversation_id=conversation.id,
-            agent_run_id=agent_run_id,
-            sequence=sequence,
-            role=draft.role.value,
-            kind=draft.kind.value,
-            text=draft.text,
-            tool_name=draft.tool_name,
-            tool_call_id=draft.tool_call_id,
-            tool_args=(
-                to_json_value(draft.tool_args) if draft.tool_args is not None else None
-            ),
-            tool_result=(
-                to_json_value(draft.tool_result)
-                if draft.tool_result is not None
-                else None
-            ),
-            message_metadata=draft.metadata,
-        )
-        self.session.add(model)
-        await self.session.flush()
-        return model.to_entity()
-
-    async def list_messages(
-        self,
-        *,
-        conversation_id: UUID,
-        before_sequence: int | None = None,
-        after_sequence: int | None = None,
-        limit: int = 100,
-    ) -> tuple[list[MessageEntity], int | None]:
-        stmt = select(MessageModel).where(
-            MessageModel.conversation_id == conversation_id
-        )
-        if before_sequence is not None:
-            stmt = stmt.where(MessageModel.sequence < before_sequence)
-        if after_sequence is not None:
-            stmt = stmt.where(MessageModel.sequence > after_sequence)
-        stmt = stmt.order_by(MessageModel.sequence.desc()).limit(limit + 1)
-        result = await self.session.execute(stmt)
-        rows = list(result.scalars())
-        has_more = len(rows) > limit
-        if has_more:
-            rows = rows[:limit]
-        next_cursor = rows[-1].sequence if has_more and rows else None
-        return [row.to_entity() for row in reversed(rows)], next_cursor
-
-    async def record_approval_decision(
-        self,
-        *,
-        conversation_id: UUID,
-        approval_id: str,
-        agent_run_id: UUID | None,
-        tool_name: str | None,
-        decision: AgentRunApprovalDecision,
-        response: JsonObject | None,
-        resolved_by_user_id: UUID,
-    ) -> bool:
-        """Persist a decision once. Returns False if already recorded."""
-        existing = await self.session.execute(
-            select(AgentApprovalDecisionModel.id).where(
-                AgentApprovalDecisionModel.conversation_id == conversation_id,
-                AgentApprovalDecisionModel.approval_id == approval_id,
-            )
-        )
-        if existing.scalar_one_or_none() is not None:
-            return False
-        self.session.add(
-            AgentApprovalDecisionModel(
-                conversation_id=conversation_id,
-                approval_id=approval_id,
-                agent_run_id=agent_run_id,
-                tool_name=tool_name,
-                decision=decision.value,
-                response=response or {},
-                resolved_by_user_id=resolved_by_user_id,
-            )
-        )
-        await self.session.flush()
-        return True
-
-    async def get_approval_decision(
-        self,
-        *,
-        conversation_id: UUID,
-        approval_id: str,
-    ) -> tuple[AgentRunApprovalDecision, JsonObject] | None:
-        result = await self.session.execute(
-            select(AgentApprovalDecisionModel).where(
-                AgentApprovalDecisionModel.conversation_id == conversation_id,
-                AgentApprovalDecisionModel.approval_id == approval_id,
-            )
-        )
-        row = result.scalar_one_or_none()
-        if row is None:
-            return None
-        response = row.response if isinstance(row.response, dict) else {}
-        return AgentRunApprovalDecision(row.decision), response
-
-    async def get_tool_call(
-        self,
-        *,
-        conversation_id: UUID,
-        tool_call_id: str,
-    ) -> MessageEntity | None:
-        """The pausing tool CALL for an approval, addressed by tool_call_id.
-
-        Looked up directly (not through a message-window scan) so a long
-        conversation can't push the original request_approval/ask_user call out
-        of view during resume reconciliation.
-        """
-        result = await self.session.execute(
-            select(MessageModel)
-            .where(
-                MessageModel.conversation_id == conversation_id,
-                MessageModel.tool_call_id == tool_call_id,
-                MessageModel.kind == MessageKind.TOOL_CALL.value,
-            )
-            .order_by(MessageModel.sequence.asc())
-            .limit(1)
-        )
-        row = result.scalar_one_or_none()
-        return row.to_entity() if row is not None else None
-
-    async def get_tool_return(
-        self,
-        *,
-        conversation_id: UUID,
-        tool_call_id: str,
-    ) -> MessageEntity | None:
-        """The synthesized tool RETURN for an approval, or None.
-
-        This is the idempotency guard for approval resume: if a return already
-        exists, the approved tool has already run and must NOT be re-executed.
-        """
-        result = await self.session.execute(
-            select(MessageModel)
-            .where(
-                MessageModel.conversation_id == conversation_id,
-                MessageModel.tool_call_id == tool_call_id,
-                MessageModel.kind == MessageKind.TOOL_RETURN.value,
-            )
-            .order_by(MessageModel.sequence.asc())
-            .limit(1)
-        )
-        row = result.scalar_one_or_none()
-        return row.to_entity() if row is not None else None
-
-    async def list_resolved_approval_ids(
-        self,
-        *,
-        conversation_id: UUID,
-    ) -> set[str]:
-        result = await self.session.execute(
-            select(AgentApprovalDecisionModel.approval_id).where(
-                AgentApprovalDecisionModel.conversation_id == conversation_id
-            )
-        )
-        return {row for row in result.scalars()}
-
-    async def finish_agent_run(
-        self,
-        *,
-        agent_run_id: UUID,
-        status: AgentRunStatus,
-        conversation_status: ConversationStatus | None = None,
-        error: str | None = None,
-        output_data: JsonValue | None = None,
-    ) -> AgentRunFinishResult | None:
-        result = await self.session.execute(
-            select(AgentRunModel).where(AgentRunModel.id == agent_run_id)
-        )
-        model = result.scalar_one_or_none()
-        if model is None:
-            return None
-
-        current_status = AgentRunStatus(model.status)
-        resolved_conversation_status = conversation_status or ConversationStatus(
-            current_status.value
-        )
-        if current_status in TERMINAL_AGENT_RUN_STATUSES:
-            return AgentRunFinishResult(
-                status=current_status,
-                conversation_status=resolved_conversation_status,
-                updated=False,
-            )
-
-        next_status = status
-        if (
-            current_status == AgentRunStatus.STOP_REQUESTED
-            and status in TERMINAL_AGENT_RUN_STATUSES
-        ):
-            next_status = AgentRunStatus.STOPPED
-
-        model.status = next_status.value
-        model.error = error
-        if output_data is not None:
-            model.output_data = output_data
-        resolved_conversation_status = conversation_status or ConversationStatus(
-            next_status.value
-        )
-        await self._update_conversation_status(
-            conversation_id=model.conversation_id,
-            status=resolved_conversation_status,
-            output_data=output_data,
-        )
-        if next_status in TERMINAL_AGENT_RUN_STATUSES:
-            model.finished_at = datetime.now(timezone.utc)
-        await self.session.flush()
-        return AgentRunFinishResult(
-            status=next_status,
-            conversation_status=resolved_conversation_status,
-            updated=True,
-        )
-
-    async def set_conversation_status(
-        self,
-        *,
-        conversation_id: UUID,
-        status: ConversationStatus,
-    ) -> None:
-        """Move a conversation's status with no run to finish.
-
-        ``finish_agent_run`` is the usual path, but a suspended conversation has
-        no active run — the run ended when the tool paused it — so cancelling its
-        wait has to set the status directly.
-        """
-        await self._update_conversation_status(
-            conversation_id=conversation_id,
-            status=status,
-        )
-
-    async def _update_conversation_status(
-        self,
-        *,
-        conversation_id: UUID,
-        status: ConversationStatus,
-        output_data: JsonValue | None = None,
-    ) -> None:
-        conversation = await self.session.get(ConversationModel, conversation_id)
-        if conversation is None:
-            return
-        conversation.status = status.value
-        if output_data is not None or status == ConversationStatus.RUNNING:
-            conversation.output_data = output_data
