+++ lemma-backend/architecture-baseline.json
@@ -1,8 +1,8 @@
 {
   "broad_catches": {
     "agent:count": 64,
-    "agent:max": 4,
-    "agent:total": 80,
+    "agent:max": 3,
+    "agent:total": 77,
     "agent_surfaces:count": 106,
     "agent_surfaces:max": 5,
     "agent_surfaces:total": 122,
@@ -43,7 +43,7 @@
   "complex_functions": {
     "agent:count": 15,
     "agent:max": 42,
-    "agent:total": 371,
+    "agent:total": 365,
     "connectors:count": 8,
     "connectors:max": 24,
     "connectors:total": 152,
@@ -83,8 +83,7 @@
   "module_cycles": [],
   "oversized_files": {
     "app/modules/agent/infrastructure/harnesses/pydantic_ai.py": 1293,
-    "app/modules/agent/infrastructure/repositories.py": 1117,
-    "app/modules/agent/services/agent_runner_service.py": 1048,
+    "app/modules/agent/services/agent_runner_service.py": 1036,
     "app/modules/agent/services/conversation_service.py": 1495,
     "app/modules/agent/services/runtime_profile_service.py": 952,
     "app/modules/agent/tools/pod/pydantic_adapter.py": 609,
